Returned arch to former structure
